refactor: service 함수 사용되는 로직에 따라 분리
